add delete and create file

+++ app/Http/Controllers/API/LegalProduct/DeleteLegalProductController.php
@@ -12,9 +12,15 @@ class DeleteLegalProductController extends Controller
 {
     public function __invoke(LegalProduct $legal_product)
     {
-        $files = $legal_product->file()->pluck('file')->toArray();
-        Storage::disk('public')->delete($files);
-        $legal_product->delete();
-        return ResponseFormatter::success(null, 'success delete legal product data');
+        if($legal_product->status == 'pending') {
+            $files = $legal_product->file()->pluck('file')->toArray();
+            Storage::disk('public')->delete($files);
+            $legal_product->delete();
+            return ResponseFormatter::success(null, 'success delete legal product data');
+        } else {
+            return ResponseFormatter::error([
+                'message' => 'cannot delete this data'
+            ], 'delete legal product failed', 422);
+        }
     }
 }
